# docs/openapi.yaml
openapi: 3.0.3
info:
  title: flowctl HTTP API
  version: "1"
  description: |
    HTTP API for [flowctl](https://github.com/cvhariharan/flowctl), a single-binary
    self-service workflow platform.

    ## Authentication

    Two mechanisms are supported. **Browser clients** authenticate via session
    cookies (`POST /login` or the OIDC flow) and must include the `X-CSRF-Token`
    header on every state-changing request. **Programmatic clients** authenticate
    via user-scoped API tokens minted in the web UI and sent in the
    `Authorization: Bearer <token>` header (token prefix `fctl_pat_`). Bearer
    requests are exempt from CSRF.

    API-token requests **cannot** create or revoke API tokens themselves, those
    endpoints require a session.

    ## Conventions

    - All resource IDs are UUIDs.
    - Timestamps are RFC 3339 (`2025-06-04T10:00:00Z`).
    - Errors use a uniform JSON envelope (see `Error` schema).
    - Pagination endpoints return `page_count` and `total_count` next to the page slice.
servers:
  - url: https://flowctl.example.com
    description: Production server (replace with your deployment URL)
  - url: http://localhost:8080
    description: Local development
security:
  - bearerAuth: []
  - cookieAuth: []
...
paths:
...
  /api/v1/permissions/check:
    post:
      operationId: CheckPermission
      tags: [Permissions]
      summary: Check a specific permission for the calling user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
              description: |
                Implementation accepts a free-form object with the resource and
                action to check. Refer to `internal/handlers/auth.go` for the
                exact shape — typically `{ "resource": "...", "action": "...", "domain": "..." }`.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "fctl_pat_<random>"
      description: |
        User-scoped API token. Mint via `POST /api/v1/users/me/api-tokens` while
        logged in via the browser. Send as `Authorization: Bearer fctl_pat_...`.
    executorBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: "fctl_<executor_name>.<signature>"
      description: |
        Executor token generated by the server for a running executor. Send as
        `Authorization: Bearer fctl_<executor_name>.<signature>`. Executors may
        also send `X-User-UUID` to associate requests with the triggering user.
    cookieAuth:
      type: apiKey
      in: cookie
      name: flowctl_session
      description: |
        Session cookie set by `POST /login` or the OIDC flow. Browser clients
        must also send `X-CSRF-Token` on state-changing requests.
...